# fix(builtin): fix linker bug when there are no third-party modules