--- a/internal/linker/test/no_npm_deps/BUILD.bazel
+++ b/internal/linker/test/no_npm_deps/BUILD.bazel
@@ -1,0 +1,11 @@
+load(":no_npm_deps.bzl", "no_npm_deps")
+
+# Tests the
+# ```
+# [link_node_modules.js] no third-party packages; mkdir node_modules at npm/node_modules
+# ```
+# path in the linker
+no_npm_deps(
+    name = "no_npm_deps_linker_test",
+    src = "input.js",
+)