--- a/internal/linker/test/no_npm_deps/input.js
+++ b/internal/linker/test/no_npm_deps/input.js
@@ -1,0 +1,1 @@
+console.log('here');